refactor: move ParseMission to parse_mission.go
parser.go now only contains the Parser struct, constructor, helpers,
and MissionContext — no parse methods.
